requirements.txt
kaggle
torch 
torchvision
google-cloud-storage
google-cloud-aiplatform
grpcio-tools